configure and start testing search
