+++ lib/storage/client.js
@@ -532,7 +532,7 @@ class Client extends CryptoConsumer {
 
   async _search(searchRequest) {
     let response = await this.authenticator.tokenFetch(
-      this.config.apiUrl + '/v2/storage/search',
+      this.config.apiUrl + '/v2/search',
       {
         method: 'POST',
         headers: {
